+++ app/src/main/java/com/example/sport_planet/data/model/mypage/MyViewHistoryModel.kt
@@ -1,4 +1,6 @@
-package com.example.sport_planet.data.model
+package com.example.sport_planet.data.model.mypage
+
+import com.example.sport_planet.data.model.BoardInfo
 
 data class MyViewHistoryModel(
     val host : Boolean,
