--- app/src/main/java/com/example/sport_planet/data/response/AddressResponse.kt
@@ -1,5 +1,0 @@
-package com.example.sport_planet.data.response
-
-data class AddressResponse(
-    val id: Long
-)
