+++ app/src/main/java/com/example/sport_planet/data/response/UserResponse.kt
@@ -1,5 +1,7 @@
 package com.example.sport_planet.data.response
 
+import com.example.sport_planet.data.response.basic.AddressResponse
+
 data class UserResponse(
     val username: String,
     val nickname: String,
