+++ app/src/main/java/com/example/sport_planet/data/response/basic/AddressResponse.kt
@@ -1,0 +1,5 @@
+package com.example.sport_planet.data.response.basic
+
+data class AddressResponse(
+    val id: Long
+)
