+++ app/src/main/java/com/example/sport_planet/presentation/login/LoginActivity.kt
@@ -5,7 +5,7 @@ import android.os.Bundle
 import android.util.Log
 import android.widget.Toast
 import com.example.sport_planet.R
-import com.example.sport_planet.data.response.LoginResponse
+import com.example.sport_planet.data.response.login.LoginResponse
 import com.example.sport_planet.databinding.ActivityLoginBinding
 import com.example.sport_planet.presentation.base.BaseActivity
 import com.example.sport_planet.presentation.main.MainActivity
