+++ app/src/main/java/com/example/sport_planet/presentation/mypage/history/adapter/FinishTabAdapter.kt
@@ -9,11 +9,11 @@ import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import androidx.recyclerview.widget.RecyclerView
 import com.example.sport_planet.R
-import com.example.sport_planet.data.model.MyViewHistoryModel
-import com.example.sport_planet.data.response.EvaluateListResponse
+import com.example.sport_planet.data.model.mypage.MyViewHistoryModel
+import com.example.sport_planet.data.response.mypage.EvaluateListResponse
 import com.example.sport_planet.databinding.ItemHistoryFinishBinding
 
-class FinishTabAdapter(private val onClickListener: (MyViewHistoryModel) -> Unit,private val dialogListener : (Long) -> Unit) :
+class FinishTabAdapter(private val onClickListener: (MyViewHistoryModel) -> Unit, private val dialogListener : (Long) -> Unit) :
     RecyclerView.Adapter<FinishTabAdapter.FinishTabViewHolder>() {
 
     val historyItem = mutableListOf<MyViewHistoryModel>()
