+++ app/src/main/java/com/example/sport_planet/presentation/mypage/history/adapter/IngTabExpandAdapter.kt
@@ -5,7 +5,7 @@ import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import androidx.recyclerview.widget.RecyclerView
 import com.example.sport_planet.R
-import com.example.sport_planet.data.model.ApplyListModel
+import com.example.sport_planet.data.model.mypage.ApplyListModel
 import com.example.sport_planet.databinding.ItemHistoryIngExpandBinding
 
 class IngTabExpandAdapter : RecyclerView.Adapter<IngTabExpandAdapter.IngTabExpandViewHolder>() {
