+++ app/src/main/java/com/example/sport_planet/presentation/mypage/scrap/ScrapViewModel.kt
@@ -2,7 +2,7 @@ package com.example.sport_planet.presentation.mypage.scrap
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import com.example.sport_planet.data.model.MyBookMarksModel
+import com.example.sport_planet.data.model.mypage.MyBookMarksModel
 import com.example.sport_planet.presentation.base.BaseViewModel
 import com.example.sport_planet.remote.RemoteDataSourceImpl
 import com.example.sport_planet.util.applySchedulers
