+++ app/src/main/java/com/example/sport_planet/presentation/profile/ExerciseDialog.kt
@@ -5,7 +5,7 @@ import android.view.View
 import androidx.core.content.ContextCompat
 import com.example.sport_planet.R
 import com.example.sport_planet.databinding.DialogExerciseBinding
-import com.example.sport_planet.data.response.ExerciseResponse
+import com.example.sport_planet.data.response.basic.ExerciseResponse
 import com.example.sport_planet.presentation.base.BaseDialogFragment
 
 class ExerciseDialog : BaseDialogFragment<DialogExerciseBinding>(R.layout.dialog_exercise) {
