+++ app/src/main/res/layout/item_history_finish.xml
@@ -9,7 +9,7 @@
 
         <variable
             name="items"
-            type="com.example.sport_planet.data.model.MyViewHistoryModel" />
+            type="com.example.sport_planet.data.model.mypage.MyViewHistoryModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
