+++ app/src/main/res/layout/item_history_finish_expand.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="items"
-            type="com.example.sport_planet.data.response.EvaluateListResponse.EvaluateListModel" />
+            type="com.example.sport_planet.data.response.mypage.EvaluateListResponse" />
     </data>
 
     <LinearLayout
