+++ app/src/main/res/layout/item_scrap.xml
@@ -9,7 +9,7 @@
 
         <variable
             name="items"
-            type="com.example.sport_planet.data.model.MyBookMarksModel" />
+            type="com.example.sport_planet.data.model.mypage.MyBookMarksModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
